Fix hardware acceleration check
